Use int64 for content_length

+++ src/http/http_response.f90
@@ -1,5 +1,5 @@
 module http_response
-    use, intrinsic :: iso_c_binding, only : c_size_t
+    use, intrinsic :: iso_fortran_env, only: int64
     use fhash, only: fhash_tbl_t, key => fhash_key
 
     implicit none
@@ -11,7 +11,7 @@ module http_response
     type :: response_type
         character(len=:), public, allocatable :: url, content, method, err_msg, header_string
         integer, public :: status_code = 0
-        integer(kind=c_size_t), public :: content_length = 0
+        integer(kind=int64), public :: content_length = 0
         logical, public :: ok = .true.
         type(fhash_tbl_t) :: header
     end type response_type
